requirements.txt
python-telegram-bot[modernauth]==20.3
SQLAlchemy==2.0.19
psycopg2-binary==2.9.6
alembic==1.11.1
python-dotenv==1.0.0
fastapi==0.100.0
uvicorn==0.22.0
sentry-sdk==1.27.0
pytest==7.4.0
